Enhance auction features on front-end and update project scaffolding
This commit includes significant enhancements to the auction features on the front-end like creating auction data and CRUD operations on auction actions. It also updates project scaffolding for better efficiency and maintainability. Further, few changes are made to keep the codebase up-to-date with dependencies, improving backward compatibility and fixing a few bugs related to type mismatch.

+++ src/SearchService/Consumers/BidPlacedConsumer.cs
@@ -18,7 +18,7 @@ public class BidPlacedConsumer: IConsumer<BidPlaced> {
         if (auction.CurrentHighBid == null || consumerContext.Message.BidStatus.Contains("Accepted") &&
             consumerContext.Message.Amount > auction.CurrentHighBid)
         {
-            auction.CurrentHighBid = (int)consumerContext.Message.Amount;
+            auction.CurrentHighBid = consumerContext.Message.Amount;
             await auction.SaveAsync();
         }
     }
